Tratamento de exceção delete User com pedido associado

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -4,11 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
-import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
+import com.educandoweb.course.resources.expections.ResourceNotFoundException;
+import com.educandoweb.course.services.exceptions.DatabaseException;
 
 @Service
 public class UserService {
@@ -19,25 +21,33 @@ public class UserService {
 	public List<User> findAll() {
 		return repository.findAll();
 	}
-	
+
 	public User findById(Long id) {
 		Optional<User> obj = repository.findById(id);
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
-		
+
 	}
-	
+
 	public User insert(User obj) {
 		return repository.save(obj);
 	}
-	
-	public void delete(long id) {
-		repository.deleteById(id);
+
+	public void delete(Long id) {
+		try {
+			if (repository.existsById(id)) {
+				repository.deleteById(id);
+			} else {
+				throw new ResourceNotFoundException(id);
+			}
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
-	
+
 	public User update(Long id, User obj) {
-		//User entity = repository.getReferenceById(id);//Original professor
-		User entity = findById(id); //Ajuste que funcionou
-		updateData (entity, obj);
+		// User entity = repository.getReferenceById(id);//Original professor
+		User entity = findById(id); // Ajuste que funcionou
+		updateData(entity, obj);
 		return repository.save(entity);
 	}
